requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
torch==2.3.0
torchvision==0.18.0
pillow
python-multipart
pytest